move config to engine, bump version
