+++ lib/hoardable/encrypted_rich_text.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Hoardable
+  # A {Hoardable} subclass of {ActionText::EncryptedRichText}.
+  class EncryptedRichText < ActionText::EncryptedRichText
+    include Model
+  end
+end
