+++ lib/hoardable/rich_text.rb
@@ -1,11 +1,8 @@
 # frozen_string_literal: true
 
 module Hoardable
+  # A {Hoardable} subclass of {ActionText::RichText}
   class RichText < ActionText::RichText
-    include Model
-  end
-
-  class EncryptedRichText < ActionText::EncryptedRichText
     include Model
   end
 end
