+++ lib/hoardable/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module Hoardable
-  VERSION = '0.10.2'
+  VERSION = '0.11.0'
 end
